# application/services/custom_services/ticket/src/models/entitymodels/severity.ts
export interface severity 
{
   created_date: { type: Date },
   created_by: { type: String },
   last_modified_by: { type: String },
   last_modified_date: { type: Date },
   name: { type: String },
   description: { type: String }
}